requirements.txt
openai
weaviate-client
